Update validasi dan cetakan

+++ src/com/org/sistempenjualan/gui/DetailPemesananForm.java
@@ -75,8 +75,8 @@ public class DetailPemesananForm extends javax.swing.JFrame {
     public void setupTable(){
         tblPemesanan.setModel(DbUtils.resultSetToTableModel(dao.setDetailPemesananTable(idPemesanan)));
         tblPemesanan.removeColumn(tblPemesanan.getColumnModel().getColumn(0));
-        tblPemesanan.removeColumn(tblPemesanan.getColumnModel().getColumn(1));
-        tblPemesanan.removeColumn(tblPemesanan.getColumnModel().getColumn(2));
+        tblPemesanan.removeColumn(tblPemesanan.getColumnModel().getColumn(0));
+        tblPemesanan.removeColumn(tblPemesanan.getColumnModel().getColumn(0));
         util.adjustColumn(tblPemesanan);
     }
     
@@ -116,12 +116,8 @@ public class DetailPemesananForm extends javax.swing.JFrame {
         btnHapus = new javax.swing.JButton();
         txtCustomer = new javax.swing.JTextField();
         cbBarang = new javax.swing.JComboBox();
-        btnHapus1 = new javax.swing.JButton();
+        btnSelesai = new javax.swing.JButton();
         mbPemesanan = new javax.swing.JMenuBar();
-        btnLogOut = new javax.swing.JMenu();
-        jMenuItem2 = new javax.swing.JMenuItem();
-        jMenu2 = new javax.swing.JMenu();
-        jMenuItem1 = new javax.swing.JMenuItem();
         todayDate = new javax.swing.JMenu();
         userSession = new javax.swing.JMenu();
 
@@ -201,26 +197,12 @@ public class DetailPemesananForm extends javax.swing.JFrame {
 
         cbBarang.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "-- Pilih Barang --" }));
 
-        btnHapus1.setText("SELESAI");
-        btnHapus1.addActionListener(new java.awt.event.ActionListener() {
+        btnSelesai.setText("SELESAI");
+        btnSelesai.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnHapus1ActionPerformed(evt);
+                btnSelesaiActionPerformed(evt);
             }
         });
-
-        btnLogOut.setText("Server");
-
-        jMenuItem2.setText("Log Out");
-        btnLogOut.add(jMenuItem2);
-
-        mbPemesanan.add(btnLogOut);
-
-        jMenu2.setText("Form");
-
-        jMenuItem1.setText("Form Surat Jalan");
-        jMenu2.add(jMenuItem1);
-
-        mbPemesanan.add(jMenu2);
 
         todayDate.setText("Date");
         mbPemesanan.add(todayDate);
@@ -265,7 +247,7 @@ public class DetailPemesananForm extends javax.swing.JFrame {
                                     .addGap(18, 18, 18)
                                     .addComponent(btnHapus, javax.swing.GroupLayout.PREFERRED_SIZE, 83, javax.swing.GroupLayout.PREFERRED_SIZE)
                                     .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                    .addComponent(btnHapus1, javax.swing.GroupLayout.PREFERRED_SIZE, 83, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(btnSelesai, javax.swing.GroupLayout.PREFERRED_SIZE, 83, javax.swing.GroupLayout.PREFERRED_SIZE)
                                     .addGap(18, 18, 18)
                                     .addComponent(btnReset, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE))
                                 .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 383, javax.swing.GroupLayout.PREFERRED_SIZE)))))
@@ -299,11 +281,12 @@ public class DetailPemesananForm extends javax.swing.JFrame {
                             .addComponent(txtJmlBarang, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                             .addComponent(jLabel2))))
                 .addGap(18, 18, 18)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnSimpan)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(btnReset)
-                    .addComponent(btnHapus)
-                    .addComponent(btnHapus1))
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(btnSimpan)
+                        .addComponent(btnHapus)
+                        .addComponent(btnSelesai)))
                 .addContainerGap(50, Short.MAX_VALUE))
         );
 
@@ -341,29 +324,34 @@ public class DetailPemesananForm extends javax.swing.JFrame {
         boolean resultUpdateBrg = false;
         try{
             stok = brgDao.getLastStok(namaBarang);
-            if(stok - jumlah < 0){
-                JOptionPane.showMessageDialog(null,"Stok "+barang+" hanya ada "+stok+", tidak bisa menambah barang!");
+            if(cbBarang.getSelectedIndex() == 0){
+                JOptionPane.showMessageDialog(rootPane,"Harus Pilih Barang Terlebih Dulu!");
+            }else if(stok - jumlah < 0){
+                JOptionPane.showMessageDialog(rootPane,"Stok "+barang+" hanya ada "+stok+", tidak bisa menambah barang!");
             }else{
-                entity.setKodeBarang(kodeBarang);
-                entity.setJumlahBarang(jumlah);
-                entity.setKeteranganPemesanan(keterangan);
-                result = dao.createDetailPemesanan(entity);
-                if(result){
-                    resultUpdateJumlah = dao.tambahJumlahPemesanan(idPemesanan, jumlah);
-                    if(resultUpdateJumlah){
-                        resultUpdateBrg = brgDao.updateStokBarang(kodeBarang, jumlah);
-                        if(resultUpdateBrg){
-                            JOptionPane.showMessageDialog(null,"Berhasil tambah barang");
-                            setupTable();
-                            clearForm();
+                int confirm = JOptionPane.showConfirmDialog(null, "Anda yakin ingin menyimpan data?","Simpan Data",JOptionPane.YES_NO_OPTION);
+                if(confirm==0){
+                    entity.setKodeBarang(kodeBarang);
+                    entity.setJumlahBarang(jumlah);
+                    entity.setKeteranganPemesanan(keterangan);
+                    result = dao.createDetailPemesanan(entity);
+                    if(result){
+                        resultUpdateJumlah = dao.tambahJumlahPemesanan(idPemesanan, jumlah);
+                        if(resultUpdateJumlah){
+                            resultUpdateBrg = brgDao.updateStokBarang(kodeBarang, jumlah);
+                            if(resultUpdateBrg){
+                                JOptionPane.showMessageDialog(null,"Berhasil tambah barang");
+                                setupTable();
+                                clearForm();
+                            }else{
+                                JOptionPane.showMessageDialog(null,"Gagal tambah barang");
+                            }
                         }else{
                             JOptionPane.showMessageDialog(null,"Gagal tambah barang");
                         }
                     }else{
                         JOptionPane.showMessageDialog(null,"Gagal tambah barang");
                     }
-                }else{
-                    JOptionPane.showMessageDialog(null,"Gagal tambah barang");
                 }
             }
         }catch(Exception e){
@@ -394,22 +382,26 @@ public class DetailPemesananForm extends javax.swing.JFrame {
         boolean resultUpdateBrg = false;
         int jumlah = Integer.parseInt(txtJmlBarang.getText());
         try{
-            resultUpdateBrg = brgDao.returnStokBarang(entity.getKodeBarang(), jumlah);
-            if(resultUpdateBrg){
-                resultUpdateJumlah = dao.kurangJumlahPemesanan(idPemesanan, jumlah);
-                if(resultUpdateJumlah){
-                    result = dao.deleteDetailPemesanan(entity.getIdDetailPemesanan());
-                    if(result){
-                        JOptionPane.showMessageDialog(null,"Berhasil hapus barang");
-                        setupTable();
+            int confirm = JOptionPane.showConfirmDialog(null, "Anda yakin ingin menghapus data?","Hapus Data",JOptionPane.YES_NO_OPTION);
+            if(confirm==0){
+                resultUpdateBrg = brgDao.returnStokBarang(entity.getKodeBarang(), jumlah);
+                if(resultUpdateBrg){
+                    resultUpdateJumlah = dao.kurangJumlahPemesanan(idPemesanan, jumlah);
+                    if(resultUpdateJumlah){
+                        result = dao.deleteDetailPemesanan(entity.getIdDetailPemesanan());
+                        if(result){
+                            JOptionPane.showMessageDialog(null,"Berhasil hapus barang");
+                            setupTable();
+                            clearForm();
+                        }else{
+                            JOptionPane.showMessageDialog(null,"Gagal hapus barang");
+                        }
                     }else{
                         JOptionPane.showMessageDialog(null,"Gagal hapus barang");
                     }
                 }else{
                     JOptionPane.showMessageDialog(null,"Gagal hapus barang");
                 }
-            }else{
-                JOptionPane.showMessageDialog(null,"Gagal hapus barang");
             }
         }catch(Exception e){
             try {
@@ -420,22 +412,21 @@ public class DetailPemesananForm extends javax.swing.JFrame {
         }
     }//GEN-LAST:event_btnHapusActionPerformed
 
-    private void btnHapus1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnHapus1ActionPerformed
-        
-    }//GEN-LAST:event_btnHapus1ActionPerformed
+    private void btnSelesaiActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSelesaiActionPerformed
+        int confirm = JOptionPane.showConfirmDialog(null, "Anda yakin ingin menyelesaikan transaksi?","Selesaikan Transaksi?",JOptionPane.YES_NO_OPTION);
+        if(confirm==0)
+        this.setVisible(false);
+        new PemesananForm(entity).setVisible(true);
+    }//GEN-LAST:event_btnSelesaiActionPerformed
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnHapus;
-    private javax.swing.JButton btnHapus1;
-    private javax.swing.JMenu btnLogOut;
     private javax.swing.JButton btnReset;
+    private javax.swing.JButton btnSelesai;
     private javax.swing.JButton btnSimpan;
     private javax.swing.JComboBox cbBarang;
     private javax.swing.JLabel jLabel2;
-    private javax.swing.JMenu jMenu2;
-    private javax.swing.JMenuItem jMenuItem1;
-    private javax.swing.JMenuItem jMenuItem2;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JScrollPane jScrollPane2;
     private javax.swing.JLabel lblBarang;
